requirements.txt
# requirements.txt
# Core dependencies
uvicorn[standard]>=0.24.0
aiofiles>=23.2.1
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0
sqlalchemy[asyncio]>=2.0.23

# Security  
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.7

# Templating
jinja2>=3.1.2

# CLI
click>=8.1.7

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
mypy>=1.7.1
